added take damage option

+++ Shoot-Arena-3D/Assets/Scripts/Infrastructure/Core/Enemies/Handlers/EnemyHealth/IEnemyHealthHandler.cs
@@ -2,6 +2,6 @@
 {
     public interface IEnemyHealthHandler
     {
-        void ReceiveDamage(int value);
+        void ReceiveDamage(float value);
     }
 }
